Robert/frontend design

+++ src/main/java/mk/napijse/web/CommentsController.java
@@ -1,10 +1,12 @@
 package mk.napijse.web;
 
 import mk.napijse.model.entities.Comment;
+import mk.napijse.model.enumerations.Role;
 import mk.napijse.model.exceptions.CommentNotFoundException;
 import mk.napijse.model.exceptions.RecipeNotFoundException;
 import mk.napijse.service.CommentService;
 import mk.napijse.service.RecipeService;
+import mk.napijse.service.UserService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -17,13 +19,17 @@ import java.security.Principal;
 
 @Controller
 public class CommentsController {
-    //test
+
     private final CommentService commentService;
     private final RecipeService recipeService;
+    private final UserService userService;
 
-    public CommentsController(CommentService commentService, RecipeService recipeService) {
+    public CommentsController(CommentService commentService,
+                              RecipeService recipeService,
+                              UserService userService) {
         this.commentService = commentService;
         this.recipeService = recipeService;
+        this.userService = userService;
     }
 
     @PostMapping("/add-comment/{recipeId}")
@@ -39,11 +45,22 @@ public class CommentsController {
     @GetMapping("/edit-comment/{recipeId}/{commentId}")
     public String getEditCommentPage(@PathVariable Long recipeId,
                                      @PathVariable Long commentId,
-                                     Model model){
+                                     Model model,
+                                     HttpServletRequest request){
+
+        String username = request.getUserPrincipal().getName();
+
         try{
-            Comment comment = this.commentService.findById(commentId);
-            model.addAttribute("comment", comment);
-            return "edit-comment";
+            if((this.commentService.findById(commentId).getCommentUser().getUsername().equals(username)) ||
+                    (this.userService.findByUsername(username).isPresent() &&
+                            this.userService.findByUsername(username).get().getRole().equals(Role.ROLE_ADMIN))){
+                Comment comment = this.commentService.findById(commentId);
+                model.addAttribute("comment", comment);
+
+                model.addAttribute("bodyContent", "edit-comment");
+                return "master-template";
+            }
+            return "redirect:/recipes/info/"+recipeId + "?error=You can't edit other user's comment";
         }catch (CommentNotFoundException exception){
             return "redirect:/recipes/info/"+recipeId + "?error=" + exception.getMessage();
         }
@@ -52,10 +69,19 @@ public class CommentsController {
     @PostMapping("/edit-comment/{recipeId}/{commentId}")
     public String editComment(@PathVariable Long recipeId,
                               @PathVariable Long commentId,
-                              @RequestParam String content){
+                              @RequestParam String content,
+                              HttpServletRequest request){
+
+        String username = request.getUserPrincipal().getName();
+
         try{
-            this.commentService.edit(recipeId, commentId, content);
-            return "redirect:/recipes/info/"+recipeId;
+            if ((this.commentService.findById(commentId).getCommentUser().getUsername().equals(username)) ||
+                    (this.userService.findByUsername(username).isPresent() &&
+                            this.userService.findByUsername(username).get().getRole().equals(Role.ROLE_ADMIN))){
+                this.commentService.edit(recipeId, commentId, content);
+                return "redirect:/recipes/info/"+recipeId;
+            }
+            return "redirect:/recipes/info/"+recipeId + "?error=You can't edit other user's comment";
         }catch (RecipeNotFoundException | CommentNotFoundException exception){
             return "redirect:/recipes/info/"+recipeId + "?error=" + exception.getMessage();
         }
@@ -63,10 +89,19 @@ public class CommentsController {
 
     @PostMapping("/delete-comment/{recipeId}/{commentId}")
     public String deleteComment(@PathVariable Long recipeId,
-                              @PathVariable Long commentId) {
+                              @PathVariable Long commentId,
+                                HttpServletRequest request) {
+
+        String username = request.getUserPrincipal().getName();
+
         try {
-            this.commentService.delete(recipeId, commentId);
-            return "redirect:/recipes/info/" + recipeId;
+            if ((this.commentService.findById(commentId).getCommentUser().getUsername().equals(username)) ||
+                    (this.userService.findByUsername(username).isPresent() &&
+                            this.userService.findByUsername(username).get().getRole().equals(Role.ROLE_ADMIN))){
+                this.commentService.delete(recipeId, commentId);
+                return "redirect:/recipes/info/" + recipeId;
+            }
+            return "redirect:/recipes/info/"+recipeId + "?error=You can't delete other user's comment";
         } catch (RecipeNotFoundException | CommentNotFoundException exception) {
             return "redirect:/recipes/info/" + recipeId + "?error=" + exception.getMessage();
         }
